Reorganized the class structure of user-related test classes -Eliminated dependencies between test methods, ensuring each method operates independently -Modeled the test classes based on the structure  used in Postman collections to promote consistency and ease of management.
