+++ src/test/java/tests/institution/adminmanagementservice/PostAdminsTest.java
@@ -1,0 +1,4 @@
+package tests.institution.adminmanagementservice;
+
+public class PostAdminsTest {
+}
