+++ src/test/java/tests/institution/adminregistrationmanagementservice/GetAdminRegistrationApplicationIDTest.java
@@ -1,0 +1,4 @@
+package tests.institution.adminregistrationmanagementservice;
+
+public class GetAdminRegistrationApplicationIDTest {
+}
