+++ src/test/java/tests/institution/adminregistrationmanagementservice/PostAdminRegistrationApplicationsTest.java
@@ -1,0 +1,4 @@
+package tests.institution.adminregistrationmanagementservice;
+
+public class PostAdminRegistrationApplicationsTest {
+}
