--- src/test/java/tests/institution/assignmentmanagement/DeleteAssignmentTest.java
@@ -1,4 +1,0 @@
-package tests.institution.assignmentmanagement;
-
-public class DeleteAssignmentTest {
-}
