--- src/test/java/tests/institution/assignmentmanagement/UpdateAssignmentServiceTest.java
@@ -1,4 +1,0 @@
-package tests.institution.assignmentmanagement;
-
-public class UpdateAssignmentServiceTest {
-}
