+++ src/test/java/tests/institution/assignmentmanagementservice/DeleteAssignmentTest.java
@@ -1,0 +1,4 @@
+package tests.institution.assignmentmanagementservice;
+
+public class DeleteAssignmentTest {
+}
