+++ src/test/java/tests/institution/assignmentmanagementservice/PostAssignmentTest.java
@@ -1,4 +1,4 @@
-package tests.institution.assignmentmanagement;
+package tests.institution.assignmentmanagementservice;
 
 import com.github.javafaker.Faker;
 import endpoints.InstitutionEndpoints;
