+++ src/test/java/tests/institution/institutionmanagerapplicationmanagementservice/GetInstitutionsSummaryTest.java
@@ -1,0 +1,4 @@
+package tests.institution.institutionmanagerapplicationmanagementservice;
+
+public class GetInstitutionsSummaryTest {
+}
