--- src/test/java/tests/institution/usermanagement/PostUsersServiceTest.java
@@ -1,4 +1,0 @@
-package tests.institution.usermanagement;
-
-public class PostUsersServiceTest {
-}
