+++ src/test/java/tests/institution/usermanagementservice/PostUsersTest.java
@@ -1,0 +1,4 @@
+package tests.institution.usermanagementservice;
+
+public class PostUsersTest {
+}
