+++ src/test/java/tests/user/userassignmentmanagementservice/PostAssignmentCompleteTest.java
@@ -1,0 +1,82 @@
+package tests.user.userassignmentmanagementservice;
+
+import endpoints.UserEndpoints;
+import io.restassured.response.Response;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+import payload.Assignment;
+import payload.Helper;
+import payload.Location;
+import payload.UserCredentials;
+
+import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.containsString;
+
+public class PostAssignmentCompleteTest {
+    UserCredentials userCredentials;
+    Logger logger = LogManager.getLogger(this.getClass());
+    Location location;
+    Assignment assignment;
+
+
+    @BeforeClass
+    public void setup() {
+        userCredentials = Helper.createNewUser();
+        location = new Location();
+        assignment=Helper.createANewAssignment();
+
+    }
+    @Test()
+    public void assignmentCompleteWhenUserHasReservedAssignment() {
+        location=Helper.generateLocationTR();
+        Helper.setSupportStatus("READY", userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.searchAssignment(location, userCredentials.getUsername(), userCredentials.getPassword());
+        Response response = UserEndpoints.completeAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        response.then()
+                .statusCode(404)
+                .body("time", notNullValue())
+                .body("httpStatus", equalTo("NOT_FOUND"))
+                .body("header", equalTo("NOT EXIST"))
+                .body("message", allOf(
+                        containsString("ASSIGNMENT NOT EXIST!"),
+                        containsString("assignmentStatus:IN_PROGRESS")))
+                .body("isSuccess", equalTo(false));
+    }
+    @Test()
+    public void assignmentComplete() {
+        logger.info("Test case UMS_13 is running");
+        location=Helper.generateLocationTR();
+        Helper.setSupportStatus("READY", userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.searchAssignment(location, userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.approveAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.startAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        Response response = UserEndpoints.completeAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        response.then()
+                .statusCode(200)
+                .body("time", notNullValue())
+                .body("httpStatus", equalTo("OK"))
+                .body("isSuccess", equalTo(true));
+    }
+    @Test()
+    public void assignmentCompleteNegative() {
+        logger.info("Test case UMS_13 is running");
+        location=Helper.generateLocationTR();
+        Helper.setSupportStatus("READY", userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.searchAssignment(location, userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.approveAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.startAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.completeAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        Response response = UserEndpoints.completeAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        response.then()
+                .statusCode(404)
+                .body("time", notNullValue())
+                .body("httpStatus", equalTo("NOT_FOUND"))
+                .body("header", equalTo("NOT EXIST"))
+                .body("message", allOf(
+                        containsString("ASSIGNMENT NOT EXIST!"),
+                        containsString("assignmentStatus:IN_PROGRESS")))
+                .body("isSuccess", equalTo(false));
+    }
+}
