+++ src/test/java/tests/user/userassignmentmanagementservice/PostAssignmentRejectTest.java
@@ -1,0 +1,76 @@
+package tests.user.userassignmentmanagementservice;
+
+import endpoints.UserEndpoints;
+import io.restassured.response.Response;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.testng.annotations.BeforeClass;
+import org.testng.annotations.Test;
+import payload.Assignment;
+import payload.Helper;
+import payload.Location;
+import payload.UserCredentials;
+
+import static org.hamcrest.Matchers.*;
+import static org.hamcrest.Matchers.containsString;
+
+public class PostAssignmentRejectTest {
+    UserCredentials userCredentials;
+    Logger logger = LogManager.getLogger(this.getClass());
+    Location location;
+    Assignment assignment;
+
+
+    @BeforeClass
+    public void setup() {
+        userCredentials = Helper.createNewUser();
+        location = new Location();
+        assignment=Helper.createANewAssignment();
+
+    }
+    @Test()
+    public void rejectAssignmentBeforeSearch() {
+        Helper.setSupportStatus("READY", userCredentials.getUsername(), userCredentials.getPassword());
+        Response response = UserEndpoints.rejectAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        response.then()
+                .statusCode(404)
+                .body("time", notNullValue())
+                .body("httpStatus", equalTo("NOT_FOUND"))
+                .body("header", equalTo("NOT EXIST"))
+                .body("message", allOf(
+                        containsString("ASSIGNMENT NOT EXIST!"),
+                        containsString("assignmentStatus:RESERVED")))
+                .body("isSuccess", equalTo(false));
+    }
+
+
+    @Test()
+    public void rejectAssignmentAfterSearch() {
+        logger.info("Test case UMS_06 is running");
+        Helper.setSupportStatus("READY", userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.searchAssignment(location, userCredentials.getUsername(), userCredentials.getPassword());
+        Response response = UserEndpoints.rejectAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        response.then()
+                .statusCode(200)
+                .body("time", notNullValue())
+                .body("httpStatus", equalTo("OK"))
+                .body("isSuccess", equalTo(true));
+    }
+    @Test()
+    public void rejectInProgressAssignment() {
+        Helper.setSupportStatus("READY", userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.searchAssignment(location, userCredentials.getUsername(), userCredentials.getPassword());
+        UserEndpoints.approveAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        Response response = UserEndpoints.rejectAssignment(userCredentials.getUsername(), userCredentials.getPassword());
+        response.then()
+                .statusCode(404)
+                .body("time", notNullValue())
+                .body("httpStatus", equalTo("NOT_FOUND"))
+                .body("header", equalTo("NOT EXIST"))
+                .body("message", allOf(
+                        containsString("ASSIGNMENT NOT EXIST!"),
+                        containsString("assignmentStatus:RESERVED")))
+                .body("isSuccess", equalTo(false));
+    }
+
+}
