--- src/test/java/tests/user/userusermanagement/GetUserDetailById.java
@@ -1,4 +1,0 @@
-package tests.user.userusermanagement;
-
-public class GetUserDetailById {
-}
